Revamp collision detection.
Every object now has a move(), collidesWith() and update() method. In
the server update loop, each object is moved, checked for collisions,
then updated (and then its changes are recorded).

If A collides with B and B collides with A, then the server will push
A in B.collisions and B in A.collisions. In their respective update()
methods, they can check if they collided with any type of object with
@collidedWith (method of ChangingObject).